move UserAnalyticsService to :library:analytics module
