+++ library/analytics/build.gradle.kts
@@ -35,6 +35,7 @@ android {
 dependencies {
     api(project(":library:base"))
     implementation(project(":library:account"))
+    implementation(project(":library:user-data"))
     implementation(project(":ui:base"))
 
     implementation(libs.firebase.messaging)
